feat: implement TripsSchedule model and update related handlers
