+++ internal/models/trips_for_location.go
@@ -12,17 +12,11 @@ type TripsForLocationData struct {
 }
 
 type TripsForLocationListEntry struct {
-	Frequency    *int64                    `json:"frequency"`
-	Schedule     *TripsForLocationSchedule `json:"schedule,omitempty"`
-	ServiceDate  int64                     `json:"serviceDate"`
-	SituationIds []string                  `json:"situationIds"`
-	TripId       string                    `json:"tripId"`
+	Frequency    *int64         `json:"frequency"`
+	Schedule     *TripsSchedule `json:"schedule,omitempty"`
+	ServiceDate  int64          `json:"serviceDate"`
+	SituationIds []string       `json:"situationIds"`
+	TripId       string         `json:"tripId"`
 }
 
-type TripsForLocationSchedule struct {
-	Frequency      *int64     `json:"frequency"`
-	NextTripId     string     `json:"nextTripId"`
-	PreviousTripId string     `json:"previousTripId"`
-	StopTimes      []StopTime `json:"stopTimes"`
-	TimeZone       string     `json:"timeZone"`
-}
+func (e TripsForLocationListEntry) GetTripId() string { return e.TripId }
